+++ bika/lims/browser/publish/emailview.py
@@ -20,6 +20,7 @@
 
 import inspect
 import itertools
+import re
 from collections import OrderedDict
 from string import Template
 
@@ -728,7 +729,7 @@ class EmailView(BrowserView):
         # CC Contacts
         cc = filter(None, map(recipient_from_contact, ar.getCCContact()))
         # CC Emails
-        cc_emails = map(lambda x: x.strip(), ar.getCCEmails().split(","))
+        cc_emails = ar.getCCEmails(as_list=True)
         cc_emails = filter(None, map(recipient_from_email, cc_emails))
 
         return to + cc + cc_emails