+++ bika/lims/browser/publish/reports_listing.py
@@ -19,6 +19,7 @@
 # Some rights reserved, see README and LICENSE.
 
 import collections
+import re
 
 from bika.lims import api
 from bika.lims import bikaMessageFactory as _BMF
@@ -242,7 +243,7 @@ class ReportsListingView(BikaListingView):
         # CC Contacts
         cc = filter(None, map(recipient_from_contact, ar.getCCContact()))
         # CC Emails
-        cc_emails = map(lambda x: x.strip(), ar.getCCEmails().split(","))
+        cc_emails = ar.getCCEmails(as_list=True)
         cc_emails = filter(None, map(recipient_from_email, cc_emails))
 
         return to + cc + cc_emails